requirements.txt
# Core
pydantic>=2.7,<3
pydantic-settings>=2.3,<3

# LLM / AI
langchain-core>=0.3,<0.4
langchain-groq>=0.2,<0.3
langchain-openai>=0.3,<0.4
langgraph>=0.2,<0.3

# Vector DB
chromadb>=0.5,<1

# Web search fallback
tavily-python>=0.5,<1

# Reranking
cohere>=5,<6

# Retry / resilience
tenacity>=8.2,<10

# API
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1

# Observability
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
opentelemetry-instrumentation-fastapi>=0.46b0
langfuse>=2.0,<3

# Ingestion
beautifulsoup4>=4.12
lxml>=5.0
httpx>=0.27

# Testing
pytest>=8
pytest-asyncio>=0.23
pytest-cov>=5
